requirements.txt
selenium==4.25.0
pytest==7.4.4
pytest-html==4.1.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0